*Corrections to satisfy Coverity. *Changed eof_min_note_distance and eof_chord_density_threshold to unsigned as there should be no scenario where it is expected or appropriate for them to be negative. *Removed an if() statement in eof_menu_catalog_add() that was always true (because start_pos and end_pos variables of the EOF_CATALOG_ENTRY haven't been signed ints for a long time). *Changed ini_settings in the EOF_SONG_TAGS structure to unsigned as there should be no scenario where it is expected or appropriate for it to be negative. *Changed lyric line preview arrays to unsigned long type as there should be no scenario where it is expected or appropriate for these values to be negative. *Changed eof_get_note_max_length() to return a signed long value since the length variable is still treated as signed throughout the code. *Various other corrections to data type in terms of whether or not they are signed, where appropriate.  The changes in this commit remove 50 signed/unsigned comparison compiler warnings (-Wextra).

+++ src/menu/note.h
@@ -241,7 +241,7 @@ int eof_menu_copy_solos_track_10(void);
 int eof_menu_copy_solos_track_11(void);
 int eof_menu_copy_solos_track_12(void);
 int eof_menu_copy_solos_track_13(void);
-int eof_menu_copy_solos_track_number(EOF_SONG *sp, int sourcetrack, int desttrack);
+int eof_menu_copy_solos_track_number(EOF_SONG *sp, unsigned long sourcetrack, unsigned long desttrack);
 	//Copies the solo phrases from the specified source track to the destination track
 
 int eof_menu_copy_sp_track_1(void);
@@ -257,7 +257,7 @@ int eof_menu_copy_sp_track_10(void);
 int eof_menu_copy_sp_track_11(void);
 int eof_menu_copy_sp_track_12(void);
 int eof_menu_copy_sp_track_13(void);
-int eof_menu_copy_sp_track_number(EOF_SONG *sp, int sourcetrack, int desttrack);
+int eof_menu_copy_sp_track_number(EOF_SONG *sp, unsigned long sourcetrack, unsigned long desttrack);
 	//Copies the star power phrases from the specified source track to the destination track
 
 int eof_menu_copy_arpeggio_track_1(void);
@@ -273,7 +273,7 @@ int eof_menu_copy_arpeggio_track_10(void);
 int eof_menu_copy_arpeggio_track_11(void);
 int eof_menu_copy_arpeggio_track_12(void);
 int eof_menu_copy_arpeggio_track_13(void);
-int eof_menu_copy_arpeggio_track_number(EOF_SONG *sp, int sourcetrack, int desttrack);
+int eof_menu_copy_arpeggio_track_number(EOF_SONG *sp, unsigned long sourcetrack, unsigned long desttrack);
 	//Copies the arpeggio phrases from the specified source track to the destination track
 
 int eof_menu_copy_trill_track_1(void);
@@ -289,7 +289,7 @@ int eof_menu_copy_trill_track_10(void);
 int eof_menu_copy_trill_track_11(void);
 int eof_menu_copy_trill_track_12(void);
 int eof_menu_copy_trill_track_13(void);
-int eof_menu_copy_trill_track_number(EOF_SONG *sp, int sourcetrack, int desttrack);
+int eof_menu_copy_trill_track_number(EOF_SONG *sp, unsigned long sourcetrack, unsigned long desttrack);
 	//Copies the trill phrases from the specified source track to the destination track
 
 int eof_menu_copy_tremolo_track_1(void);
@@ -305,7 +305,7 @@ int eof_menu_copy_tremolo_track_10(void);
 int eof_menu_copy_tremolo_track_11(void);
 int eof_menu_copy_tremolo_track_12(void);
 int eof_menu_copy_tremolo_track_13(void);
-int eof_menu_copy_tremolo_track_number(EOF_SONG *sp, int sourcetrack, int desttrack);
+int eof_menu_copy_tremolo_track_number(EOF_SONG *sp, unsigned long sourcetrack, unsigned long desttrack);
 	//Copies the tremolo phrases from the specified source track to the destination track
 
 int eof_feedback_mode_update_note_selection(void);
